Fix: pdf to docx

+++ src/main/java/com/example/fileconverter/FileConverterService.java
@@ -1,39 +1,32 @@
 package com.example.fileconverter;
-
 import org.springframework.stereotype.Service;
-
 import java.io.*;
 
 @Service
 public class FileConverterService {
-
-    public void convertDocxToPdf(InputStream inputStream, OutputStream outputStream) throws IOException {
+    public void convertFile(InputStream inputStream, OutputStream outputStream, String originalFilename, String targetFormat) throws IOException {
         File tempInputFile = null;
         File tempOutputFile = null;
 
         try {
-            // 1. Create temporary files
-            tempInputFile = File.createTempFile("upload-", ".docx");
-            tempOutputFile = File.createTempFile("converted-", ".pdf");
+            String inputExtension = "";
+            int i = originalFilename.lastIndexOf('.');
+            if (i > 0) {inputExtension = originalFilename.substring(i + 1).toLowerCase();}
+
+            tempInputFile = File.createTempFile("upload-", "." + inputExtension);
+            tempOutputFile = File.createTempFile("converted-", "." + targetFormat);
 
 
             try (FileOutputStream fileOut = new FileOutputStream(tempInputFile)) {inputStream.transferTo(fileOut);}
-
-            DocxToPdfConverter converter = new DocxToPdfConverter();
+            FileConverter converter = ConverterFactory.getConverter(inputExtension, targetFormat);
             try {converter.convert(tempInputFile, tempOutputFile.getAbsolutePath());}
-            catch (Exception e) {throw new IOException("Conversion failed inside DocxToPdfConverter", e);}
-
+            catch (Exception e) {throw new IOException("Conversion failed: " + e.getMessage(), e);}
             try (FileInputStream fileIn = new FileInputStream(tempOutputFile)) {fileIn.transferTo(outputStream);}
 
         }
         finally {
-
-            if (tempInputFile != null && tempInputFile.exists()) {
-                tempInputFile.delete();
-            }
-            if (tempOutputFile != null && tempOutputFile.exists()) {
-                tempOutputFile.delete();
-            }
+            if (tempInputFile != null && tempInputFile.exists()) tempInputFile.delete();
+            if (tempOutputFile != null && tempOutputFile.exists()) tempOutputFile.delete();
         }
     }
 }
